Fix cloud attrs setting on standard metrics in AKS scenarios.
