+++ sdk/monitor/monitor-opentelemetry/src/metrics/utils.ts
@@ -5,9 +5,6 @@ import type { Attributes } from "@opentelemetry/api";
 import { SpanStatusCode } from "@opentelemetry/api";
 import type { ReadableSpan } from "@opentelemetry/sdk-trace-base";
 import {
-  SEMRESATTRS_SERVICE_NAME,
-  SEMRESATTRS_SERVICE_NAMESPACE,
-  SEMRESATTRS_SERVICE_INSTANCE_ID,
   SEMATTRS_PEER_SERVICE,
   SEMATTRS_NET_PEER_NAME,
   SEMATTRS_NET_HOST_PORT,
@@ -24,6 +21,16 @@ import {
   DBSYSTEMVALUES_OTHER_SQL,
   DBSYSTEMVALUES_HSQLDB,
   DBSYSTEMVALUES_H2,
+  SEMRESATTRS_K8S_DEPLOYMENT_NAME,
+  SEMRESATTRS_K8S_REPLICASET_NAME,
+  SEMRESATTRS_K8S_STATEFULSET_NAME,
+  SEMRESATTRS_K8S_JOB_NAME,
+  SEMRESATTRS_K8S_CRONJOB_NAME,
+  SEMRESATTRS_K8S_DAEMONSET_NAME,
+  SEMRESATTRS_K8S_POD_NAME,
+  SEMRESATTRS_SERVICE_INSTANCE_ID,
+  SEMRESATTRS_SERVICE_NAME,
+  SEMRESATTRS_SERVICE_NAMESPACE,
 } from "@opentelemetry/semantic-conventions";
 import type {
   MetricDependencyDimensions,
@@ -80,18 +87,8 @@ export function getBaseDimensions(resource: Resource): StandardMetricBaseDimensi
   const dimensions: StandardMetricBaseDimensions = {};
   dimensions.IsAutocollected = "True";
   if (resource) {
-    const spanResourceAttributes = resource.attributes;
-    const serviceName = spanResourceAttributes[SEMRESATTRS_SERVICE_NAME];
-    const serviceNamespace = spanResourceAttributes[SEMRESATTRS_SERVICE_NAMESPACE];
-    if (serviceName) {
-      if (serviceNamespace) {
-        dimensions.cloudRoleName = `${serviceNamespace}.${serviceName}`;
-      } else {
-        dimensions.cloudRoleName = String(serviceName);
-      }
-    }
-    const serviceInstanceId = spanResourceAttributes[SEMRESATTRS_SERVICE_INSTANCE_ID];
-    dimensions.cloudRoleInstance = String(serviceInstanceId);
+    dimensions.cloudRoleName = getCloudRole(resource);
+    dimensions.cloudRoleInstance = getCloudRoleInstance(resource);
   }
   return dimensions;
 }
@@ -195,4 +192,80 @@ export function getProcessorTimeNormalized(
   numCpus = numCpus === 0 ? 1 : numCpus;
 
   return (usageDifMs / elapsedTimeMs / numCpus) * 100;
+}
+
+/**
+ * Gets the cloud role name based on the resource attributes
+ * @param resource Resource object from OpenTelemetry
+ * @returns cloud role name
+ * @internal
+ */
+export function getCloudRole(resource: Resource): string {
+  let cloudRole = "";
+  // Service attributes
+  const serviceName = resource.attributes[SEMRESATTRS_SERVICE_NAME];
+  const serviceNamespace = resource.attributes[SEMRESATTRS_SERVICE_NAMESPACE];
+  if (serviceName) {
+    // Custom Service name provided by customer is highest precedence
+    if (!String(serviceName).startsWith("unknown_service")) {
+      if (serviceNamespace) {
+        return `${serviceNamespace}.${serviceName}`;
+      } else {
+        return String(serviceName);
+      }
+    } else {
+      // Service attributes will be only used if K8S attributes are not present
+      if (serviceNamespace) {
+        cloudRole = `${serviceNamespace}.${serviceName}`;
+      } else {
+        cloudRole = String(serviceName);
+      }
+    }
+  }
+  // Kubernetes attributes should take precedence
+  const kubernetesDeploymentName = resource.attributes[SEMRESATTRS_K8S_DEPLOYMENT_NAME];
+  if (kubernetesDeploymentName) {
+    return String(kubernetesDeploymentName);
+  }
+  const kuberneteReplicasetName = resource.attributes[SEMRESATTRS_K8S_REPLICASET_NAME];
+  if (kuberneteReplicasetName) {
+    return String(kuberneteReplicasetName);
+  }
+  const kubernetesStatefulSetName = resource.attributes[SEMRESATTRS_K8S_STATEFULSET_NAME];
+  if (kubernetesStatefulSetName) {
+    return String(kubernetesStatefulSetName);
+  }
+  const kubernetesJobName = resource.attributes[SEMRESATTRS_K8S_JOB_NAME];
+  if (kubernetesJobName) {
+    return String(kubernetesJobName);
+  }
+  const kubernetesCronjobName = resource.attributes[SEMRESATTRS_K8S_CRONJOB_NAME];
+  if (kubernetesCronjobName) {
+    return String(kubernetesCronjobName);
+  }
+  const kubernetesDaemonsetName = resource.attributes[SEMRESATTRS_K8S_DAEMONSET_NAME];
+  if (kubernetesDaemonsetName) {
+    return String(kubernetesDaemonsetName);
+  }
+  return cloudRole;
+}
+
+/**
+ * Gets the cloud role instance based on the resource attributes
+ * @param resource Resource object from OpenTelemetry
+ * @returns cloud role instance
+ */
+export function getCloudRoleInstance(resource: Resource): string {
+  // Kubernetes attributes should take precedence
+  const kubernetesPodName = resource.attributes[SEMRESATTRS_K8S_POD_NAME];
+  if (kubernetesPodName) {
+    return String(kubernetesPodName);
+  }
+  // Service attributes
+  const serviceInstanceId = resource.attributes[SEMRESATTRS_SERVICE_INSTANCE_ID];
+  if (serviceInstanceId) {
+    return String(serviceInstanceId);
+  }
+  // Default
+  return os && os.hostname();
 }
